requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-allauth==0.58.2
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-filter==23.5
django-mptt==0.16.0
reportlab==4.0.7
pandas==2.1.4
numpy==1.26.0
matplotlib==3.8.2
openpyxl==3.1.2
python-dateutil==2.8.2
requests==2.31.0
bcrypt==4.1.2
cryptography==41.0.7
django-encrypted-model-fields==0.6.5
django-auditlog==2.3.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
black==23.12.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
pre-commit==3.6.0
